+++ flash/image/segmentation/backbones.py
@@ -18,10 +18,15 @@ from flash.core.utilities.imports import _TORCHVISION_AVAILABLE
 from flash.image.backbones import catch_url_error
 
 if _TORCHVISION_AVAILABLE:
-    from torchvision.models import mobilenetv3, resnet
+    from torchvision.models import resnet
+
+    try:
+        from torchvision.models import mobilenetv3
+        MOBILENET_MODELS = ["mobilenet_v3_large"]
+    except ImportError:
+        MOBILENET_MODELS = []
 
 RESNET_MODELS = ["resnet50", "resnet101"]
-MOBILENET_MODELS = ["mobilenet_v3_large"]
 
 SEMANTIC_SEGMENTATION_BACKBONES = FlashRegistry("backbones")